Logo block hidden when sidebar is collapsed
The "Go to GitHub organization, Julia powered" block of the sidebar was still rendered behind the page content when the sidebar was hidden (eg, on mobile). It looked like just a few dots were randomly hovering because the text and most of the logo image where white rendered on white. Changing the position from "fixed" to "absolute" hides this block when the sidebar is hidden.